trial_02-room_database: Refactor of datastore
