+++ trial_02/app/src/main/java/cz/matee/nemect/trial_02/core/database/datastore/profile/ProfileScreenState.kt
@@ -1,4 +1,4 @@
-package cz.matee.nemect.trial_02.core.database.profile
+package cz.matee.nemect.trial_02.core.database.datastore.profile
 
 import cz.matee.nemect.trial_02.presentation.ui.theme.value.DarkMode
