+++ trial_02/app/src/main/java/cz/matee/nemect/trial_02/core/di/CoreModule.kt
@@ -1,16 +1,14 @@
 package cz.matee.nemect.trial_02.core.di
 
-import android.app.Application
 import cz.matee.nemect.trial_02.App
 import cz.matee.nemect.trial_02.core.database.DB
-import cz.matee.nemect.trial_02.core.database.profile.ProfileDataSource
+import cz.matee.nemect.trial_02.core.database.datastore.profile.ProfileDataSource
 import cz.matee.nemect.trial_02.navigation.NavigationManager
 import cz.matee.nemect.trial_02.presentation.screens.home.HomeVM
 import cz.matee.nemect.trial_02.presentation.screens.home.dictionary.DictionaryScreenVM
 import cz.matee.nemect.trial_02.presentation.screens.language.LanguageVM
 import cz.matee.nemect.trial_02.presentation.screens.settings.SettingsVM
 import cz.matee.nemect.trial_02.presentation.screens.home.test.TestVM
-import org.koin.android.ext.koin.androidContext
 
 import org.koin.androidx.viewmodel.dsl.viewModelOf
 import org.koin.dsl.module
