+++ trial_02/app/src/main/java/cz/matee/nemect/trial_02/presentation/screens/settings/SettingsVM.kt
@@ -1,12 +1,10 @@
 package cz.matee.nemect.trial_02.presentation.screens.settings
 
-import android.util.Log
-import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.mutableStateOf
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
-import cz.matee.nemect.trial_02.core.database.profile.ProfileDataSource
-import cz.matee.nemect.trial_02.core.database.profile.ProfileScreenState
+import cz.matee.nemect.trial_02.core.database.datastore.profile.ProfileDataSource
+import cz.matee.nemect.trial_02.core.database.datastore.profile.ProfileScreenState
 import cz.matee.nemect.trial_02.presentation.ui.theme.value.DarkMode
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
